Integrate GraphFiltering class into FilterExtractor class since it was not clear if they have unique responsibilities
