+++ src/main/java/de/uhd/ifi/se/decision/management/jira/filtering/FilterExtractor.java
@@ -6,13 +6,18 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.atlassian.jira.issue.Issue;
 import com.atlassian.jira.user.ApplicationUser;
 
 import de.uhd.ifi.se.decision.management.jira.filtering.impl.FilterSettingsImpl;
+import de.uhd.ifi.se.decision.management.jira.filtering.impl.JiraQueryHandlerImpl;
 import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeElement;
 import de.uhd.ifi.se.decision.management.jira.model.Graph;
+import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
 import de.uhd.ifi.se.decision.management.jira.model.impl.GraphImpl;
 import de.uhd.ifi.se.decision.management.jira.model.impl.GraphImplFiltered;
+import de.uhd.ifi.se.decision.management.jira.model.text.PartOfJiraIssueText;
+import de.uhd.ifi.se.decision.management.jira.persistence.JiraIssueTextPersistenceManager;
 
 /**
  * Extracts the Element search Items from the JQL and SearchString
@@ -22,6 +27,11 @@ public class FilterExtractor {
 	private ApplicationUser user;
 	private FilterSettings filterSettings;
 	private List<DecisionKnowledgeElement> decisionKnowledgeElements;
+	private JiraQueryHandler queryHandler;
+
+	public JiraQueryHandler getQueryHandler() {
+		return queryHandler;
+	}
 
 	public FilterExtractor(String projectKey, ApplicationUser user, String filterString) {
 		if (!isValidInput(projectKey, user, filterString)) {
@@ -29,16 +39,11 @@ public class FilterExtractor {
 		}
 		this.filterSettings = new FilterSettingsImpl(projectKey, filterString);
 		this.user = user;
-		GraphFiltering filter;
 		if (!(filterString.matches("\\?jql=(.)+")) || (filterString.matches("\\?filter=(.)+"))) {
 			filterSettings.setSearchString("asdf§filter=-4");
-		}
-		filter = new GraphFiltering(filterSettings, user);
-		filter.getQueryHandler().getJiraIssuesFromQuery();
-		this.decisionKnowledgeElements = filter.getAllElementsMatchingQuery();
-		if (this.decisionKnowledgeElements == null) {
-			this.decisionKnowledgeElements = new ArrayList<>();
-		}
+		}		
+		this.queryHandler = new JiraQueryHandlerImpl(user, projectKey, filterString);
+		this.decisionKnowledgeElements = getAllElementsMatchingQuery();
 	}
 
 	private boolean isValidInput(String projectKey, ApplicationUser user, String filterString) {
@@ -62,10 +67,10 @@ public class FilterExtractor {
 			return;
 		}
 		this.user = user;
-		this.filterSettings = filterSettings;
-		GraphFiltering filter = new GraphFiltering(filterSettings, user);
-		//filter.produceResultsWithAdditionalFilters(filterSettings.getSearchString());
-		this.decisionKnowledgeElements = filter.getAllElementsMatchingQuery();
+		this.filterSettings = filterSettings;		
+		this.queryHandler = new JiraQueryHandlerImpl(user, filterSettings.getProjectKey(),
+				filterSettings.getSearchString());
+		this.decisionKnowledgeElements = getAllElementsMatchingQuery();
 	}
 
 	private boolean isValidInput(ApplicationUser user, FilterSettings filterData) {
@@ -114,7 +119,7 @@ public class FilterExtractor {
 				// if not get the connected tree
 				String currentElementKey = current.getKey();
 				filterSettings.setSearchString(linkedQueryNotEmpty);
-				List<DecisionKnowledgeElement> filteredElements = getElementsInGraph(user, filterSettings,
+				List<DecisionKnowledgeElement> filteredElements = this.getElementsInGraph(user, filterSettings,
 						currentElementKey);
 				// add each element to the list
 				addedElements.addAll(filteredElements);
@@ -124,21 +129,69 @@ public class FilterExtractor {
 		}
 		return elementsQueryLinked;
 	}
-	
+
 	/**
 	 * Used for the export of decision knowledge
 	 */
-	private static List<DecisionKnowledgeElement> getElementsInGraph(ApplicationUser user, FilterSettings filterData,
+	private List<DecisionKnowledgeElement> getElementsInGraph(ApplicationUser user, FilterSettings filterData,
 			String elementKey) {
 		Graph graph;
 		if ((filterData.getSearchString().matches("\\?jql=(.)+"))
 				|| (filterData.getSearchString().matches("\\?filter=(.)+"))) {
-			GraphFiltering filter = new GraphFiltering(filterData, user);
-			filter.getQueryHandler().getJiraIssuesFromQuery();
-			graph = new GraphImplFiltered(filterData.getProjectKey(), elementKey, filter);
+			graph = new GraphImplFiltered(filterData.getProjectKey(), elementKey, this);
 		} else {
 			graph = new GraphImpl(filterData.getProjectKey(), elementKey);
 		}
 		return graph.getAllElements();
 	}
+
+	public List<DecisionKnowledgeElement> getAllElementsMatchingQuery() {
+		List<Issue> jiraIssues = queryHandler.getJiraIssuesFromQuery();
+		List<DecisionKnowledgeElement> results = new ArrayList<DecisionKnowledgeElement>();
+		if (jiraIssues == null) {
+			return results;
+		}
+		for (Issue currentIssue : jiraIssues) {
+			List<DecisionKnowledgeElement> elements = JiraIssueTextPersistenceManager
+					.getElementsForIssue(currentIssue.getId(), filterSettings.getProjectKey());
+			for (DecisionKnowledgeElement currentElement : elements) {
+				if (!results.contains(currentElement) && currentElement instanceof PartOfJiraIssueText
+						&& checkIfJiraTextMatchesFilter(currentElement)) {
+					results.add(currentElement);
+				}
+			}
+		}
+
+		return results;
+	}
+
+	private boolean checkIfJiraTextMatchesFilter(DecisionKnowledgeElement element) {
+		if (filterSettings.getCreatedEarliest() > 0
+				&& (element).getCreated().getTime() < filterSettings.getCreatedEarliest()) {
+			return false;
+		}
+		if (filterSettings.getCreatedLatest() > 0
+				&& (element).getCreated().getTime() > filterSettings.getCreatedLatest()) {
+			return false;
+		}
+
+		if (element.getType().equals(KnowledgeType.PRO) || element.getType().equals(KnowledgeType.CON)) {
+			if (!filterSettings.getNamesOfSelectedJiraIssueTypes().contains(KnowledgeType.ARGUMENT.toString())) {
+				return false;
+			}
+		} else if (!filterSettings.getNamesOfSelectedJiraIssueTypes().contains(element.getTypeAsString())) {
+			return false;
+		}
+
+		return true;
+	}
+
+	public FilterSettings getFilterSettings() {
+		return this.filterSettings;
+	}
+
+	public ApplicationUser getUser() {
+		return user;
+	}
+
 }
